refactor[#523] : type 폴더 types 폴더명으로 변경
