+++ src/entities/profile/api/getProfile.ts
@@ -2,7 +2,7 @@ import { skipToken, useQuery } from "@tanstack/react-query";
 import { apiClient, HttpError } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
 import { PROFILE_END_POINT, profileQueryKey } from "../constants";
-import type { ProfileInfo } from "../type/server";
+import type { ProfileInfo } from "../types/server";
 
 interface UseGetProfileParams {
   nickname: string | null;
