+++ src/features/auth/api/postAddPetInfo.ts
@@ -1,11 +1,11 @@
 import { useNavigate } from "react-router-dom";
 import { useMutation } from "@tanstack/react-query";
-import { PetInfo } from "@/entities/profile/type/server";
+import { PetInfo } from "@/entities/profile/types/server";
 import { apiClient } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store/auth";
 import { useRouteHistoryStore } from "@/shared/store/history";
 import { SIGN_UP_END_POINT } from "../constants";
-import type { SignUpResponse } from "../type/server";
+import type { SignUpResponse } from "../types/server";
 
 interface PostPetInfoRequest extends Omit<PetInfo, "profile" | "petId"> {
   image: File | null;
