+++ src/widgets/follow/ui/followingUserItem.tsx
@@ -8,7 +8,7 @@ import { FollowingToggle } from "@/features/follow/ui";
 import { ProfileLink } from "@/features/profile/ui";
 import { EmptyProfileImage, ProfileImage } from "@/entities/profile/ui";
 import { useNicknameParams } from "@/shared/lib";
-import type { FollowItemProps } from "../type/client";
+import type { FollowItemProps } from "../types/client";
 
 type FollowingUserItemProps = FollowItemProps & { isFollowing: boolean };
