+++ src/widgets/profile/ui/profileOverview.tsx
@@ -1,5 +1,5 @@
 import { FollowingToggle } from "@/features/follow/ui";
-import type { PetInfo, ProfileInfo } from "@/entities/profile/type/server";
+import type { PetInfo, ProfileInfo } from "@/entities/profile/types/server";
 import {
   EmptyProfileImage,
   PetDescriptionText,
